Persist appointments to localStorage if it exists
This mocks localStorage if it doesn't exist rather than conditionally
use it if it exists. This approach means that the exact APIs used are
documented and moves the conditional check to a single place. It means
that all client code can safely assume that persistence is working, even
if it isn't actually being persisted (eg in the tests).

This persistence is implemented in middleware so that it could easily be
extended to work with other actions. If this were to happen, the actions
would probably need to change to have some meta data about what should
be stored at each point. This is preferable to having the logic in the
reducer so that the reducer is pure.